// src/interfaces/visitation-questions.interface.ts
export interface VisitationQuestions {
  tg_id: string;
  visitation_id: string;
  date_logged: string;
  agreement_flag: number;
  no_agreement_reason: string;
  fo_name: string;
  statement_flag: number;
  deliver_flag: number;
  compliance_flag: number;
  plan_flag: number;
  frequency: string;
  frequency_count: string;
  minimum_amount: string;
  completion_date: string;
  next_date: string;
  next_amount: string;
  staff_id: string;
  imei: string;
  app_version: string;
}